Add per-host metrics
Under the `hackney_host` Telemetry key, host-based metrics are now
exposed. The hostnames themselves are tagged as Telemetry metadata.

Note that the measurement value of the worker is now flushed after
every call to `report()`, which means that the results sent to
telemetry are now effectively binned to the report interval (if any).
This is done to prevent a growing worker state in the case of
connections to many hosts.